Добавляет в проект валидацию
