+++ js/showing-big-photos.js
@@ -1,4 +1,6 @@
 import {COMMENTS_PART} from './const/const.js';
+import {publishPhotos, usersPhotoList} from './viewing-photos.js';
+import {isEscapeKey} from './utils/helpers.js';
 
 const bigPicture = document.querySelector('.big-picture');
 const bigImg = document.querySelector('.big-picture__img').querySelector('img');
@@ -51,4 +53,37 @@ const showBigPhoto = (photo) => {
   });
 };
 
-export {showBigPhoto, bigPicture, commentsLoader};
+const bigPictureCancel = bigPicture.querySelector('.big-picture__cancel');
+
+const closeBigPicture = () => {
+  bigPicture.classList.add('hidden');
+  commentsLoader.classList.remove('hidden');
+};
+
+usersPhotoList.addEventListener('click', (evt) => {
+  const child = evt.target.closest('[data-id]');
+
+  if (!child) {
+    return;
+  }
+
+  evt.preventDefault();
+  const photo = publishPhotos.find((element) => element.id === Number(child.dataset.id));
+
+  if (photo !== undefined) {
+    showBigPhoto(photo);
+  }
+
+  bigPictureCancel.addEventListener('click', () => {
+    evt.preventDefault();
+    closeBigPicture();
+  });
+
+  document.addEventListener('keydown', (key) => {
+    key.preventDefault();
+
+    if (isEscapeKey(key)) {
+      closeBigPicture();
+    }
+  });
+});
